cluster-ui: rename `/databasesV2` -> `/pages/databases/`
This commit creates a `/pages` dir in the root src
directory and moves `/databasesV2` to live in it, under
`/databases/`. This is to support a move to a file structure
that more closely encapsulates the DOM tree structure.

Epic: none

Release note: None
